Scar code refactorization
Separate the code to ease the inclusion of different providers
