+++ aws/clients/awslambda.py
@@ -1,0 +1,177 @@
+# SCAR - Serverless Container-aware ARchitectures
+# Copyright (C) 2011 - GRyCAP - Universitat Politecnica de Valencia
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from .aws import AWS
+from .resourcegroups import ResourceGroups
+from botocore.exceptions import ClientError
+from botocore.vendored.requests.exceptions import ReadTimeout
+import logging
+import utils.functionutils as utils
+import uuid
+
+class Lambda(AWS):
+    '''A low-level client representing aws Lambda.
+    https://boto3.readthedocs.io/en/latest/reference/services/lambda.htmll'''    
+    
+    def __init__(self, region=None):
+        super().__init__('lambda', region)
+    
+    def create_function_name(self, image_id_or_path):
+        parsed_id_or_path = image_id_or_path.replace('/', ',,,').replace(':', ',,,').replace('.', ',,,').split(',,,')
+        name = 'scar-%s' % '-'.join(parsed_id_or_path)
+        i = 1
+        while self.find_function_name(name):
+            name = 'scar-%s-%s' % ('-'.join(parsed_id_or_path), str(i))
+            i += 1
+        return name    
+    
+    def find_function_name(self, function_name):
+        try:
+            paginator = self.client.get_paginator('list_functions')
+            for functions in paginator.paginate():
+                for lfunction in functions['Functions']:
+                    if function_name == lfunction['FunctionName']:
+                        return True
+            return False
+        except ClientError as ce:
+            print("Error listing the lambda functions")
+            logging.error("Error listing the lambda functions: %s" % ce)
+            utils.finish_failed_execution()
+    
+    def check_function_name_not_exists(self, function_name):
+        if not self.find_function_name(function_name):
+            print("Function '%s' doesn't exist." % function_name)
+            logging.error("Function '%s' doesn't exist." % function_name)
+            utils.finish_failed_execution()
+    
+    def check_function_name_exists(self, function_name):
+        if self.find_function_name(function_name):
+            print("Function name '%s' already used." % function_name)
+            logging.error ("Function name '%s' already used." % function_name)
+            utils.finish_failed_execution()
+    
+    def update_function_timeout(self, function_name, timeout):
+        try:
+            self.client.update_function_configuration(FunctionName=function_name,
+                                                                   Timeout=self.check_time(timeout))
+        except ClientError as ce:
+            print("Error updating lambda function timeout")
+            logging.error("Error updating lambda function timeout: %s" % ce)
+    
+    def update_function_memory(self, function_name, memory):
+        try:
+            self.client.update_function_configuration(FunctionName=function_name,
+                                                                   MemorySize=memory)
+        except ClientError as ce:
+            print("Error updating lambda function memory")
+            logging.error("Error updating lambda function memory: %s" % ce)
+            
+    def create_function(self, aws_lambda): 
+        try:
+            logging.info("Creating lambda function.")
+            response = self.client.create_function(FunctionName=aws_lambda.name,
+                                                     Runtime=aws_lambda.runtime,
+                                                     Role=aws_lambda.role,
+                                                     Handler=aws_lambda.handler,
+                                                     Code=aws_lambda.code,
+                                                     Environment=aws_lambda.environment,
+                                                     Description=aws_lambda.description,
+                                                     Timeout=aws_lambda.time,
+                                                     MemorySize=aws_lambda.memory,
+                                                     Tags=aws_lambda.tags)
+            aws_lambda.function_arn = response['FunctionArn']
+            return response
+        except ClientError as ce:
+            print("Error creating lambda function")
+            logging.error("Error creating lambda function: %s" % ce)        
+        
+    def get_function_environment_variables(self, function_name):
+        return self.client.get_function(FunctionName=function_name)['Configuration']['Environment']
+    
+    def update_function_env_variables(self, function_name, env_vars):
+        try:
+            # Retrieve the global variables already defined
+            lambda_env_variables = self.get_function_environment_variables(function_name)
+            self.parse_environment_variables(lambda_env_variables, env_vars)
+            self.client.update_function_configuration(FunctionName=function_name,
+                                                                    Environment=lambda_env_variables)
+        except ClientError as ce:
+            print("Error updating the environment variables of the lambda function")
+            logging.error("Error updating the environment variables of the lambda function: %s" % ce)
+    
+    def add_lambda_permissions(self, lambda_name, bucket_name):
+        try:
+            self.client.add_permission(FunctionName=lambda_name,
+                                             StatementId=str(uuid.uuid4()),
+                                             Action="lambda:InvokeFunction",
+                                             Principal="s3.amazonaws.com",
+                                             SourceArn='arn:aws:s3:::%s' % bucket_name
+                                            )
+        except ClientError as ce:
+            print("Error setting lambda permissions")
+            logging.error("Error setting lambda permissions: %s" % ce)
+    
+    def get_function_info_by_arn(self, function_arn):
+        try:
+            return self.client.get_function(FunctionName=function_arn)
+        except ClientError as ce:
+            print("Error getting function info by arn")
+            logging.error("Error getting function info by arn: %s" % ce)
+            
+    def get_all_functions(self):
+        function_list = []
+        # Get the filtered resources from aws
+        functions_arn_list = ResourceGroups().get_lambda_functions_arn_list()
+        try:
+            for function_arn in functions_arn_list:
+                function_info = self.get_function_info_by_arn(function_arn)
+                function_list.append(function_info)
+        except ClientError as ce:
+            print("Error getting all functions")
+            logging.error("Error getting all functions: %s" % ce)
+        return function_list
+    
+    def delete_lambda_function(self, function_name):
+        try:
+            # Delete the lambda function
+            return self.client.delete_function(FunctionName=function_name)
+        except ClientError as ce:
+            print("Error deleting the lambda function")
+            logging.error("Error deleting the lambda function: %s" % ce)
+    
+    def delete_all_resources(self, aws_lambda):
+        lambda_functions = self.get_all_functions()
+        for function in lambda_functions:
+            self.delete_resources(function['Configuration']['FunctionName'], aws_lambda.output)
+            
+    def invoke_lambda_function(self, aws_lambda):
+        response = {}
+        try:
+            response = self.client.invoke(FunctionName=aws_lambda.name,
+                                                InvocationType=aws_lambda.invocation_type,
+                                                LogType=aws_lambda.log_type,
+                                                Payload=aws_lambda.payload)
+        except ClientError as ce:
+            print("Error invoking lambda function")
+            logging.error("Error invoking lambda function: %s" % ce)
+            utils.finish_failed_execution()
+    
+        except ReadTimeout as rt:
+            print("Timeout reading connection pool")
+            logging.error("Timeout reading connection pool: %s" % rt)
+            utils.finish_failed_execution()
+        return response                                        
+
